+++ setup.cfg
@@ -32,3 +32,4 @@ dev =
     tox
     pytest
     coverage
+    pytest-vcr
