+++ tests/pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 log_cli=true
-log_level=DEBUG
+log_level=INFO
 log_format="%(asctime)s [%(levelname)8s]: %(message)s [%(name)s]"
